move CheckDiskSpace out of robot/packages into disk_usage_shared.go
