+++ robot/packages/cloud_package_manager.go
@@ -555,7 +555,7 @@ func (m *cloudManager) downloadFileWithChecksum(
 			}
 		}
 		required := remaining + diskusage.MinFreeBytes
-		if _, err := CheckDiskSpace(m.logger, downloadPath, "package download", required,
+		if _, err := diskusage.CheckDiskSpace(m.logger, downloadPath, "package download", required,
 			"content_size", rutils.FormatBytes(uint64(resp.ContentLength))); err != nil {
 			return "", "", err
 		}
