+++ utils/diskusage/disk_usage_shared.go
@@ -1,10 +1,12 @@
 package diskusage
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
 
+	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/utils"
 )
 
@@ -76,9 +78,60 @@ func EnoughFreeSpace(path string, minBytes uint64) (enough bool, available uint6
 	return usage.AvailableBytes >= minBytes, usage.AvailableBytes, nil
 }
 
-// nearestExistingDir walks up from path until it finds something that exists,
-// returning that ancestor. If no ancestor exists (e.g. an empty path), it returns
-// path unchanged and lets the subsequent Statfs surface the error.
+// ErrInsufficientDiskSpace is returned by CheckDiskSpace when blocking is on and the volume is
+// low. Callers use errors.Is to tell a disk-space refusal from other failures (e.g. a corrupt
+// archive) and surface an accurate message.
+var ErrInsufficientDiskSpace = errors.New("not enough free disk space")
+
+// EnoughFreeSpace reports whether the volume holding path has at least minBytes
+// available. It is a package var so tests can inject a low-space result without
+// having to actually fill a disk.
+var EnoughFreeSpaceFunc = EnoughFreeSpace
+
+// blockingEnabled reports whether low-space conditions should refuse the operation (download,
+// local copy, or unpack). Default (unset) is false: low-space is logged but the operation
+// proceeds (log-only). See utils.ViamEnableDiskSpaceBlockEnvVar.
+func blockingEnabled() bool {
+	return utils.GetenvBool(utils.ViamEnableDiskSpaceBlockEnvVar, false)
+}
+
+// CheckDiskSpace checks whether the volume holding path has required bytes free. It returns
+// low=true whenever space is low. When blocking is enabled via ViamEnableDiskSpaceBlockEnvVar it
+// returns an error refusing the op (the caller logs it, so CheckDiskSpace stays quiet to avoid
+// double-logging the same reason every cycle); otherwise it logs a warning and returns nil so the
+// op proceeds (log-only). A failed check is logged and treated as "proceed" so a broken statfs
+// never blocks installs. desc names the op in logs/errors; extraFields extend the warning.
+func CheckDiskSpace(logger logging.Logger, path, desc string, required uint64, extraFields ...any) (low bool, err error) {
+	enough, available, err := EnoughFreeSpaceFunc(path, required)
+	if err != nil {
+		logger.Warnw("could not check free disk space; proceeding",
+			append([]any{"desc", desc, "path", path, "error", err}, extraFields...)...)
+		return false, nil
+	}
+	if enough {
+		return false, nil
+	}
+	if !blockingEnabled() {
+		// Log-only: the op proceeds and returns no error, so this warning is the only signal
+		// that space is low.
+		logger.Warnw("not enough free disk space",
+			append([]any{
+				"desc", desc, "path", path,
+				"available", utils.FormatBytes(available),
+				"required", utils.FormatBytes(required),
+				"blocking", false,
+			}, extraFields...)...)
+		return true, nil
+	}
+	// Blocking: don't warn here — the returned error carries the same detail and is logged by the
+	// caller (cloud_package_manager.go and local_package_manager.go both log the install error),
+	// so warning too would double-log the same reason every sync cycle.
+	return true, fmt.Errorf("%w for %s: %s available, %s required",
+		ErrInsufficientDiskSpace, desc, utils.FormatBytes(available), utils.FormatBytes(required))
+}
+
+// nearestExistingDir walks up from path until it finds an existing directory,
+// returning that ancestor. Non-directories are skipped
 func nearestExistingDir(path string) string {
 	for path != "" {
 		if info, err := os.Stat(path); err == nil && info.IsDir() {
